chore: offerPrice double로 타입 변경

+++ mooney/src/main/java/org/example/dto/OfferDto.java
@@ -16,7 +16,7 @@ public class OfferDto {
     // 종목코드
     private String stockCode;
     // 호가 (주문가)
-    private int offerPrice;
+    private double offerPrice;
     // 주문량
     private int offerCnt;
     // 매도 or 매수 여부
